changed repositories

+++ gym/src/main/java/org/edu/fpm/gym/entity/TrainingType.java
@@ -9,15 +9,10 @@ import lombok.Setter;
 @Setter
 @Table(name = "training_type")
 public class TrainingType {
-//    YOGA,
-//    PILATES,
-//    AEROBICS,
-//    WORKOUT,
-//    CARDIO;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_training_type", nullable = false)
-    private int id;
+    private Long id;
 
     @Column(name = "training_type_name", nullable = false, length = 50)
     private String trainingTypeName;
